feat: Send collaboration invitation email when adding a collaborator
The code changes in this commit add the functionality to send an invitation email to a collaborator when adding them to a notebook. This feature allows users to invite others to collaborate on a notebook. The `send_invite/2` function is implemented in the `Cursif.Notebooks` module, which sends an email to the specified collaborator for the given notebook.