feat: Add email template for collaboration invitation
This commit adds a new email template, "invite.html.eex", which is used to send collaboration invitation emails to users. The template includes a header, content section with a personalized message, and a button to join the notebook. It also includes a message signature and email signature at the end. This template will be used when inviting collaborators to join a notebook on Cursif.